+++ src/infi/clickhouse_orm/system_models.py
@@ -19,8 +19,8 @@ class SystemPart(Model):
     """
     OPERATIONS = frozenset({'DETACH', 'DROP', 'ATTACH', 'FREEZE', 'FETCH'})
 
-    readonly = True
-    system = True
+    _readonly = True
+    _system = True
 
     database = StringField()  # Name of the database where the table that this part belongs to is located.
     table = StringField()  # Name of the table that this part belongs to.